added  authentication to /graph endpoints
